+++ CHANGELOG.md
@@ -71,7 +71,7 @@
 - Added support for `Series.case_when` except when condition or replacement is callable.
 - Added documentation pages for `Index` and its APIs.
 - Added support for `DataFrame.assign`.
-- Added support for `DataFrame.pivot`.
+- Added support for `DataFrame.pivot` and `pd.pivot`.
 
 #### Bug Fixes